AI Tutor V2
I programmed streaming with the AI output so that the user doesn't have to wait a long time for something to show up. I also added a bunch of the options that we will eventually see on our GUI in the terminal in order for me to take AI testing to the next level. I will get rid of it later obviously but for now I'm using it to test AI prompting.