Advanced Modification Features
Added a `modify()` method to ModelBase to give the ability to modify messages and instructions
Added an `InstructionsManager` class to handle instructions and modifications to them in a more automated and modular way.
Changed `TutorGPT` to accommodate the `InstructionsManager` class
Added a `quizMode()` function in `TutorGPT` to handle the quiz prompt separately. Changed the names of methods in the `TutorGPT` class.